feat: implemented embed and detect for dct in mid values
The values where to embed the watermark are simply the sorted
frequencies of the dct (descending) + MID_FREQ_START.
Also, the watermark is converted from (0,1) to (-1,1) because we want to
keep to add more difference in case of watermark=0, but multiplying the
coefficent by 0, we delete it and then only add alpa. By using -1 we
keep its value but only change sign, so we get a bigger difference.
Also we've deleted the sign removal from the coefficent.

It seems to work pretty well, but the detecting is really sensible to
mid/high frequencies attacks like blur or median.
We need to find a way to embed also in the low frequencies because these
ones are the less affected by the two attacks.