feat: dwt -> dct embedding
The embedding works using the LL frequencies of DWT and then applying
DCT on it.

It's pretty solid on blur, compression... attacks that performs some
sort of low pass filter.

The FPR is pretty low, the THRESHOLD can be as low as 0.5, but in order to
have some margin I've changed to 0.6.

The main file contains a lot of utilities to test the embedding,
simply uncomment the functions in the main functions and tweak the
params.