fix: workflow for the challenge
Now all the functions are called from the main.py file.
During the challenge we will embed our images, so in the main.py file we
will run the following code:

``` python
    images_to_embed = [ f"{INPUT_DIR}/0000.bmp", f"{INPUT_DIR}/0001.bmp", f"{INPUT_DIR}/0002.bmp", ]
    group_name = "ACME"
    embedding_for_challenge(images_to_embed, group_name)
```

This code will embed the images passed as argument and will store them
in `OUTPUT/GROUPNAME/` with the following naming
`GROUPNAME_IMAGENAME.bmp`.

After that, the attack phase will use the attack_workflow function
```python
attack_workflow(60, INPUT_DIR, OUTPUT_DIR)
```

The function will look in the `INPUT_DIR` searching for all the images in the
following format: `IMAGENAME.bmp` and `GROUPNAME_IMAGENAME.bmp`.
Then it will attack every image using randomized attacks and after that
it will store the output in `OUTPUT_DIR/GROUPNAME/IMAGENAME/`. Inside
this folder there will be the top 5 results and a .csv file containing
the report of the attacks